bridgesec_data_transformer/requirements.txt
django
pymongo
mongoengine
djangorestframework
drf-yasg
requests
django-environ
colorlog
djangorestframework-simplejwt
dnspython
debugpy
celery
redis
pika